fixing review comments
