+++ ocs_ci/ocs/perfresult.py
@@ -158,3 +158,29 @@ class PerfResult:
         res_link = f"http://{self.server}:{self.port}/{self.new_index}/"
         res_link += f"_search?q=uuid:{self.uuid}"
         return res_link
+
+
+class ResultsAnalyse(PerfResult):
+    """
+    This class generates results for all tests as one unit
+    and saves them to an elastic search server on the cluster
+
+    """
+
+    def __init__(self, uuid, crd, full_log_path, index_name):
+        """
+        Initialize the object by reading some of the data from the CRD file and
+        by connecting to the ES server and read all results from it.
+
+        Args:
+            uuid (str): the unique uid of the test
+            crd (dict): dictionary with test parameters - the test yaml file
+                        that modify it in the test itself.
+            full_log_path (str): the path of the results files to be found
+            index_name (str): index name in ES
+        """
+        super(ResultsAnalyse, self).__init__(uuid, crd)
+        self.new_index = index_name
+        self.full_log_path = full_log_path
+        # make sure we have connection to the elastic search server
+        self.es_connect()
