+++ tests/e2e/performance/csi_tests/test_pvc_bulk_clone_performance.py
@@ -16,35 +16,9 @@ from ocs_ci.helpers.helpers import get_full_test_logs_path
 from ocs_ci.ocs import constants, scale_lib
 from ocs_ci.ocs.resources import pvc, pod
 from ocs_ci.ocs.resources.objectconfigfile import ObjectConfFile
-from ocs_ci.ocs.perfresult import PerfResult
+from ocs_ci.ocs.perfresult import ResultsAnalyse
 
 log = logging.getLogger(__name__)
-
-
-class ResultsAnalyse(PerfResult):
-    """
-    This class generates results for all tests as one unit
-    and saves them to an elastic search server on the cluster
-
-    """
-
-    def __init__(self, uuid, crd, full_log_path):
-        """
-        Initialize the object by reading some of the data from the CRD file and
-        by connecting to the ES server and read all results from it.
-
-        Args:
-            uuid (str): the unique uid of the test
-            crd (dict): dictionary with test parameters - the test yaml file
-                        that modify it in the test itself.
-            full_log_path (str): the path of the results files to be found
-
-        """
-        super(ResultsAnalyse, self).__init__(uuid, crd)
-        self.new_index = "bulk_clone_perf_fullres"
-        self.full_log_path = full_log_path
-        # make sure we have connection to the elastic search server
-        self.es_connect()
 
 
 @performance
@@ -85,7 +59,7 @@ class TestBulkCloneCreation(PASTest):
         Initialize the full results object which will send to the ES server
 
         Args:
-            full_results (obj): an empty FIOResultsAnalyse object
+            full_results (obj): an FIOResultsAnalyse object
 
         Returns:
             FIOResultsAnalyse (obj): the input object fill with data
@@ -247,7 +221,12 @@ class TestBulkCloneCreation(PASTest):
 
             # Initialize the results doc file.
             full_results = self.init_full_results(
-                ResultsAnalyse(self.uuid, self.crd_data, self.full_log_path)
+                ResultsAnalyse(
+                    self.uuid,
+                    self.crd_data,
+                    self.full_log_path,
+                    "bulk_clone_perf_fullres",
+                )
             )
 
             full_results.add_key("interface", self.interface)
@@ -256,6 +235,7 @@ class TestBulkCloneCreation(PASTest):
             full_results.add_key("bulk_creation_time", total_time)
             full_results.add_key("data_size(MB)", total_files_size)
             full_results.add_key("speed", speed)
+            full_results.add_key("es_results_link", full_results.results_link())
 
             # Write the test results into the ES server
             full_results.es_write()
